backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.3.0
chromadb>=0.4.0
langchain>=0.0.350
langchain-openai>=0.0.2
tiktoken>=0.5.0
